refactor: apply checkstyle (headers and formats)
