+++ spring-cloud-gateway-core/src/main/java/org/springframework/cloud/gateway/filter/factory/SpringCloudCircuitBreakerFilterFactory.java
@@ -47,6 +47,7 @@ import static org.springframework.cloud.gateway.support.ServerWebExchangeUtils.r
 public abstract class SpringCloudCircuitBreakerFilterFactory extends
 		AbstractGatewayFilterFactory<SpringCloudCircuitBreakerFilterFactory.Config> {
 
+	/** CircuitBreaker component name. */
 	public static final String NAME = "CircuitBreaker";
 
 	private ReactiveCircuitBreakerFactory reactiveCircuitBreakerFactory;
