+++ spring-cloud-gateway-core/src/test/java/org/springframework/cloud/gateway/config/conditional/DisableBuiltInGlobalFiltersTests.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.gateway.config;
+package org.springframework.cloud.gateway.config.conditional;
 
 import java.util.List;
